initialized new ttObject from legacy format moved utility functions to a separate file removed unnecessary/unused content to a new file
